Keep window open until the user closes it
Instead of the hack where we wait for user input from the standard
input to prevent the window from immediately closing, we now use
the function provided by the `show_image` library for this purpose.
This prevents the situation where the user can close the window
without the program actually being stopped.