Implement circle rasterization
This very simple trigonometry-based circle rasterization algorithm
unfortunately proves to be very prone to aliasing. A better alternative
might be a circle-variant of the Bresenham algorithm (midpoint circle
algorithm) or a trigonometry-based algorithm that iterates around the
circle's circumference in small segments (effectively oversampling the
pixels) and calculates the x and y coordinates as cosine and sine,
respectively.